refactor: extract websocket handlers into domain modules
Split the 2357-line websocket-handler.ts into focused modules:
- handlers/types.ts: HandlerContext, ConnectionState, helpers
- handlers/browser-handlers.ts: Directory and file operations
- handlers/search-handlers.ts: File and content search
- handlers/widget-handlers.ts: Widget computation and editing
- handlers/home-handlers.ts: Notes, goals, tasks, inspiration

Introduced HandlerContext pattern and requireVault() helper to reduce
repetition across handlers. Main file now focuses on message routing,
vault/session lifecycle, and SDK streaming.
